looks good, moving on to tests
